requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.3
Django==6.0.1
django-allauth==65.13.1
django-cf-turnstile==0.1.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-modeltranslation==0.19.19
django-nested-admin==4.1.6
docker==7.1.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
jwt==1.4.0
kombu==5.6.2
packaging==25.0
prompt_toolkit==3.0.52
psycopg2==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-monkey-business==1.1.0
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
